Bugfix: data > maxint (#36945)
* fixed and tested

* compile remove

* fix recording

* test enhancement

+++ sdk/storage/Azure.Storage.DataMovement/src/StreamToUriJobPart.cs
@@ -270,8 +270,8 @@ namespace Azure.Storage.DataMovement
                     {
                         slicedStream = await GetOffsetPartitionInternal(
                             stream,
-                            (int)0,
-                            (int)blockSize,
+                            0L,
+                            blockSize,
                             UploadArrayPool,
                             _cancellationToken).ConfigureAwait(false);
                         await _destinationResource.WriteFromStreamAsync(
@@ -340,8 +340,8 @@ namespace Azure.Storage.DataMovement
                 {
                     slicedStream = await GetOffsetPartitionInternal(
                         stream,
-                        (int)offset,
-                        (int)blockLength,
+                        offset,
+                        blockLength,
                         UploadArrayPool,
                         _cancellationToken).ConfigureAwait(false);
                     await _destinationResource.WriteFromStreamAsync(
@@ -427,8 +427,8 @@ namespace Azure.Storage.DataMovement
         /// </returns>
         private static async Task<Stream> GetOffsetPartitionInternal(
             Stream stream,
-            int offset,
-            int length,
+            long offset,
+            long length,
             ArrayPool<byte> arrayPool,
             CancellationToken cancellationToken)
         {
